implement trigger configs

+++ src/spell/modules/spell/build.sh
@@ -11,8 +11,11 @@ marine build --release
 mkdir -p artifacts
 cp target/wasm32-wasi/release/spell.wasm artifacts/
 
-# download SQLite 3 to use in tests
-curl -L https://github.com/fluencelabs/sqlite/releases/download/v0.15.0_w/sqlite3.wasm -o artifacts/sqlite3.wasm
+if [ ! -f "artifacts/sqlite3.wasm" ]; then
+  # download SQLite 3
+  curl -L https://github.com/fluencelabs/sqlite/releases/download/v0.15.0_w/sqlite3.wasm -o artifacts/sqlite3.wasm
+fi
+
 
 cd ./artifacts
 tar --exclude="spell.tar.gz" -f spell.tar.gz -zcv ./*
